Re-organization and clean up
Since the LogUnitEntry is part of the LogUnitServer internal
representation, I grouped them together. Also, I added a address
field in LogUnitEntry because that information was only retained
in the datacache map.
